+++ packages/builder/builder-rspack-provider/src/types/hooks.ts
@@ -1,15 +1,8 @@
-import type {
-  NodeEnv,
-  BuilderTarget,
-  Stats,
-  MultiStats,
-} from '@modern-js/builder-shared';
-import type { BuilderConfig } from './config';
+import type { NodeEnv, BuilderTarget } from '@modern-js/builder-shared';
 import type {
   RspackConfig,
   RspackRuleSet,
   RspackPluginInstance,
-  Compiler,
 } from './rspack';
 
 export type ModifyRspackConfigUtils = {
@@ -34,36 +27,3 @@ export type ModifyRspackConfigFn = (
   config: RspackConfig,
   utils: ModifyRspackConfigUtils,
 ) => Promise<RspackConfig | void> | RspackConfig | void;
-
-export type ModifyBuilderConfigFn = (
-  config: BuilderConfig,
-) => Promise<BuilderConfig | void> | BuilderConfig | void;
-
-export type OnBeforeBuildFn = (params: {
-  bundlerConfigs?: RspackConfig[];
-}) => Promise<void> | void;
-
-export type OnAfterBuildFn = (params: {
-  stats?: Stats | MultiStats;
-}) => Promise<void> | void;
-
-export type OnDevCompileDoneFn = (params: {
-  isFirstCompile: boolean;
-}) => Promise<void> | void;
-
-export type OnBeforeCreateCompilerFn = (params: {
-  bundlerConfigs: RspackConfig[];
-}) => Promise<void> | void;
-
-// TODO: support MultiCompiler
-export type OnAfterCreateCompilerFn = (params: {
-  compiler: Compiler;
-}) => Promise<void> | void;
-
-export type OnBeforeStartDevServerFn = () => Promise<void> | void;
-
-export type OnAfterStartDevServerFn = (params: {
-  port: number;
-}) => Promise<void> | void;
-
-export type OnExitFn = () => void;