+++ packages/builder/builder-webpack-provider/src/core/initHooks.ts
@@ -1,29 +1,37 @@
-import { createAsyncHook } from '@modern-js/builder-shared';
-import type {
+import {
+  createAsyncHook,
   OnExitFn,
   OnAfterBuildFn,
   OnBeforeBuildFn,
   OnDevCompileDoneFn,
-  ModifyWebpackChainFn,
-  ModifyWebpackConfigFn,
   ModifyBuilderConfigFn,
-  OnAfterCreateCompilerFn,
-  OnBeforeCreateCompilerFn,
   OnAfterStartDevServerFn,
   OnBeforeStartDevServerFn,
+  OnAfterCreateCompilerFn,
+  OnBeforeCreateCompilerFn,
+} from '@modern-js/builder-shared';
+import type { Compiler, MultiCompiler } from 'webpack';
+import type {
+  WebpackConfig,
+  BuilderConfig,
+  ModifyWebpackChainFn,
+  ModifyWebpackConfigFn,
 } from '../types';
 
 export function initHooks() {
   return {
     onExitHook: createAsyncHook<OnExitFn>(),
     onAfterBuildHook: createAsyncHook<OnAfterBuildFn>(),
-    onBeforeBuildHook: createAsyncHook<OnBeforeBuildFn>(),
+    onBeforeBuildHook: createAsyncHook<OnBeforeBuildFn<WebpackConfig>>(),
     onDevCompileDoneHook: createAsyncHook<OnDevCompileDoneFn>(),
     modifyWebpackChainHook: createAsyncHook<ModifyWebpackChainFn>(),
     modifyWebpackConfigHook: createAsyncHook<ModifyWebpackConfigFn>(),
-    modifyBuilderConfigHook: createAsyncHook<ModifyBuilderConfigFn>(),
-    onAfterCreateCompilerHook: createAsyncHook<OnAfterCreateCompilerFn>(),
-    onBeforeCreateCompilerHook: createAsyncHook<OnBeforeCreateCompilerFn>(),
+    modifyBuilderConfigHook:
+      createAsyncHook<ModifyBuilderConfigFn<BuilderConfig>>(),
+    onAfterCreateCompilerHook:
+      createAsyncHook<OnAfterCreateCompilerFn<Compiler | MultiCompiler>>(),
+    onBeforeCreateCompilerHook:
+      createAsyncHook<OnBeforeCreateCompilerFn<WebpackConfig>>(),
     onAfterStartDevServerHook: createAsyncHook<OnAfterStartDevServerFn>(),
     onBeforeStartDevServerHook: createAsyncHook<OnBeforeStartDevServerFn>(),
   };