+++ packages/builder/builder/src/plugins/index.ts
@@ -1,0 +1,4 @@
+export const plugins = {
+  cleanOutput: () => import('./cleanOutput').then(m => m.PluginCleanOutput()),
+  startUrl: () => import('./startUrl').then(m => m.PluginStartUrl()),
+};